Major refactor, performance improvements and more (#7)

+++ src/edge.rs
@@ -3,5 +3,5 @@ use serde::{Serialize, Deserialize};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Edge {
     pub to: usize,
-    pub cost: f64,
+    pub cost: f32,
 }
